refactor: moved foreach content into separate files.
